Move getFacilitiesRtDataById to the rt file
